[add] Add the Youtube Component to App Component
