+++ src/App.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import Timer from './Components/Timer-Component.js';
-import CookieClicker from './Components/Cookie-Clicker.js'
+import CookieClicker from './Components/Cookie-Clicker.js';
+import Youtube from './Components/Youtube-Component.js';
 import './App.css';
 import axios from 'axios';
 
@@ -8,35 +9,11 @@ class App extends Component {
   constructor(props, context) {
     super(props, context);
     this.state = {
-      returnedVideos: []
+    
     }
   }
 
   componentDidMount() {
-    this.retrieveYoutubeVideos();
-  }
-
-  retrieveYoutubeVideos = (that = this, userSearchQuery) => {
-    axios.post('/retrieveYoutubeFeed', {
-      // you should replace this query with the userSearchQuery's value you will provide
-      searchQuery: 'yogscast'
-    })
-    .then(function (response) {
-      let newReturnedVideos = that.buildNewVideosArray(response.data);
-      that.setState({
-        returnedVideos: newReturnedVideos,
-      });
-    })
-    .catch(function (error) {
-      console.log(error);
-    });
-  }
-
-  buildNewVideosArray = (responseData) => {
-    return responseData.map(function(video){
-      let videoURL = "http://www.youtube.com/embed/" + video.id.videoId;
-      return <iframe className="ytplayer" type="text/html" width="640" height="360" src={videoURL} frameBorder="0"></iframe>;
-    });
   }
 
   getUsers = () => {
@@ -81,10 +58,7 @@ class App extends Component {
       <div className="App">
         <Timer/>
         <CookieClicker/>
-        <div>{this.state.returnedVideos}</div>
-        <button style={{'position': 'absolute', 'top': '100px', 'zIndex': 4}} onClick={() => {this.addUserToUsersTable()}}>{'Add user'}</button>
-        <button style={{'position': 'absolute', 'top': '200px', 'zIndex': 4}} onClick={() => {this.verifyUser()}}>{'Verify user'}</button>
-        <button style={{'position': 'absolute', 'top': '300px', 'zIndex': 4}} onClick={() => {this.getUsers()}}>{'Get Users'}</button>        
+        <Youtube />
       </div>
     );
   }
